PerformanceTests/jmxedit/requirements.txt
pyyaml==5.4.1
lxml==4.6.3
selenium==4.0.0
